requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.4.0=pypi_0
asgiref=3.4.1=pypi_0
bcrypt=3.2.0=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pypi_0
contextlib2=21.6.0=pypi_0
contextvars=2.4=pypi_0
cryptography=36.0.1=pypi_0
dataclasses=0.8=pypi_0
ecdsa=0.17.0=pypi_0
fastapi=0.70.1=pypi_0
h11=0.12.0=pypi_0
httptools=0.3.0=pypi_0
idna=3.3=pypi_0
immutables=0.16=pypi_0
importlib-metadata=4.8.3=pypi_0
passlib=1.7.4=pypi_0
pip=21.2.2=py36haa95532_0
pyasn1=0.4.8=pypi_0
pycparser=2.21=pypi_0
pydantic=1.8.2=pypi_0
python=3.6.13=h3758d61_0
python-dotenv=0.19.2=pypi_0
python-jose=3.3.0=pypi_0
python-multipart=0.0.5=pypi_0
pyyaml=6.0=pypi_0
rsa=4.8=pypi_0
setuptools=58.0.4=py36haa95532_0
six=1.16.0=pypi_0
sniffio=1.2.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
starlette=0.16.0=pypi_0
typing-extensions=4.0.1=pypi_0
uvicorn=0.16.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchgod=0.7=pypi_0
websockets=9.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py36h7fe50ca_0
zipp=3.6.0=pypi_0
